ADIDNS forest and non-domain fixes
Did some forest testing and noticed that the objectCategory node attribute required the parent domain instead of the targeted child domain. Added a Forest parameter to New-ADIDNSNode for controlling objectCategory.

Fixed a bug that was causing [System.DirectoryServices.ActiveDirectory.Domain]::GetCurrentDomain() to run on non-domain attached systems with New-ADIDNSNode and Enable-ADIDNSNode.